requirements.txt
appdirs==1.4.3
apturl==0.5.2
attrs==19.3.0
awscli==1.18.69
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.1
blinker==1.4
botocore==1.16.19
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
duplicity==0.8.12.0
easydict==1.9
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.15.2
Jinja2==2.10.1
jmespath==0.9.4
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
kazam==1.5.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libusb1==1.9.2
libvirt-python==6.1.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
nbconvert==5.6.1
nbformat==5.0.4
ndeflib==0.3.3
netifaces==0.10.4
nfcpy==1.0.3
notebook==6.0.3
numpy==1.20.2
oauthlib==3.1.0
olefile==0.46
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.6.1
pyasn1==0.4.2
PyAudio==0.2.11
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pydensecrf==1.0rc2
pyDes==2.0.1
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.5
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.0
s3transfer==0.3.3
scipy==1.6.2
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.14.0
slack-sdk==3.5.1
ssh-import-id==5.10
systemd-python==234
tensorboard-logger==0.1.0
terminado==0.8.2
testpath==0.4.4
torchfile==0.1.0
tornado==5.1.1
tqdm==4.59.0
traitlets==4.3.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
visdom==0.1.8.9
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.58.0
widgetsnbextension==2.0.0
xkit==0.0.0
zipp==1.0.0
